domd/kernel: Enable CONFIG_MEDIA_XEN_FRONTEND
Enable the PV camera frontend for shared video use-case. The idea is
to share the stream from the same hardware device between two domains
(DomD and DomA). In order to run the frontend in DomD we will need
the hotplug support for PV camera since the backend is also in DomD.

Signed-off-by: Oleksandr Tyshchenko <oleksandr_tyshchenko@epam.com>
Signed-off-by: Mykyta Poturai <mykyta_poturai@epam.com>