prod-cockpit-rcar, *lvds-domd.dts: reduce the number of  the overriding
The idea is to introduce the name convention for
- domain config file
- domain device tree
- xen device tree

The approach allows to minimize the amount of the overrides in the case if:
- Machine has been changed
- Display has been changed
- SOC has been changed

The name convention
- domain config file: [domd/doma]-{MACHINE}.cfg
- domain device tree: %{SOC_FAMILY}-%{MACHINE}-%{XT_MACHINE_EXT}-%{XT_DISPL_MODE}-[domd/doma].dtb
- xen device tree: %{SOC_FAMILY}-%{MACHINE}-%{XT_MACHINE_EXT}-xen.dtb

MACHINE_EXT: the extension to identify the sort of teh machine.
For example, king fisher (kf) has 8g and 4g configuration, so, MACHINE_EXT has values:
4x2-kf and 4-kf correspondently.

File r8r8a77951-h3ulcb-4x2g-kf-lvds-domd.dts has been renamed to r8a77951-h3ulcb-4x2g-kf-lvds-domd.dts to follow the new convention.

Signed-off-by: Ihor Usyk <ihor_usyk@epam.com>